requirements.txt
streamlit
streamlit-cropper
torch
torchvision
pandas
numpy
matplotlib
Pillow
requests
tqdm
grad-cam
